chore: move hook tests the sub folder
